+++ bin/ncdns/src/services/openapi.rs
@@ -30,6 +30,7 @@ impl utoipa::Modify for VersionModifier {
 #[derive(OpenApi)]
 #[openapi(
   paths(
+    super::health,
     rule::apply_rule,
     rule::remove_rule,
   ),
@@ -38,6 +39,7 @@ impl utoipa::Modify for VersionModifier {
     DnsEntry,
   )),
   tags(
+    (name = "Health", description = "Health check endpoint."),
     (name = "Rules", description = "Rules management endpoints."),
   ),
   modifiers(&VersionModifier),