+++ scripts/install_dev_image.sh
@@ -11,6 +11,4 @@ echo "Downloading container images.."
 docker pull docker.io/cockroachdb/cockroach:v24.3.8
 docker pull ghcr.io/next-hat/metrsd:0.5.4
 docker pull ghcr.io/next-hat/nanocl-get-started:latest
-docker pull ghcr.io/next-hat/nanocl-dev:dev
-docker buildx build --load -t ndns:dev -f ./bin/ndns/Dockerfile .
-docker buildx build --load -t nproxy:dev -f ./bin/nproxy/Dockerfile .
+docker buildx build --load -t ghcr.io/next-hat/nanocl-dev:dev -f ./scripts/dev.Dockerfile .